notmuch: check for long run during db free
`nm_db_release(...)` should protect against freeing a long run database.
Prior to this commit, it was up to the programmer to ensure that a long
run database was not released.

The commit moves the check into `nm_db_release(...)` and removes the
checks elsewhere to reduce unnecessary, repeated checks.