imap: reduce complexity of imap_exec
imap_exec flag IMAP_CMD_FAIL_OK is the only one that configure the
function return instead of the function behavior.

Also returned value of the function depends on the flag make hard to
understand what happen after imap_exec() is run.

This change removes this flags, instead we always returns the detailed
returns code, the caller have to deal with it.

Return code are always IMAP_EXEC_SUCCESS/IMAP_EXEC_ERROR/IMAP_EXEC_FATAL and
depends on if the command succeed, failure or if the imap connection
fail.