notmuch: delegate msg_open_new to maildir backend
Notmuch delegates several functions to the maildir backend since they
are connected. This commit delegates the notmuch msg_open_new() to the
maildir backend's implementation.

The primary benefit is being able to delete attachments from emails.
Previously, if an attachment was marked for deletion then it was
impossible to exit a virtual mailbox since a write operation could not
complete. This bug has been in Neomutt for a while.

Fixes #673