config: add 'C_' prefix
Each config variable has a global variable backing it.
The name of the global is the MixedCase transform of the config name.
While this is useful for remembering their names, it makes some
variables hard to search for, giving many false positives.

- Copy
- From
- Help
- Move
- Sort

Prefixing the globals with `C_` prevents this problem.