Change struct Mailbox->realpath to be const char *
struct Mailbox->path is a fixed array (which will be converted to a struct Buffer in
the next commit).  This is needed to call mutt_expand_path().

However, struct Mailbox->realpath has no such need, and so it is a bit
wasteful (not to mention not big enough) to store as such.

Co-authored-by: Richard Russon <rich@flatcap.org>