Do not rely on idn2 to provide an idn1 compat layer
Do not rely on idn2 to provide a compat layer to idn1, as Ubuntu doesn't
provide it. Instead, really code address/idna.c to use the different
APIs provided by the two versions of libidn. Systems with an old version
of idn2 which does not provide the needed functionality will fail at
configure time now, and users of such systems will see an error message
with instructions how to build with libidn1. While at it, polish a bit
the auto.def sections for idn and idn2.