Disable state messages for attachments when forwarding
First, option(OptViewAttach) is only set for (s->flags & DISPLAY), so that
check can be ignored.

With that removed, all this is doing is printing
  [-- This is an attachment ...]
with no context in the middle of non-display rendering for the case
$honor_disposition is set with a handler.

This makes no sense and I believe is a logic error.  Perhaps this
never took effect, but with $forward_attachments it now affects
forwarding in that case.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/8ae8394f83b207d8b1c411e8b40cc3945840b71c
Co-authored-by: Richard Russon <rich@flatcap.org>