agentcrypt: password encryption with ssh-agent
This commit adds a new 'agentcrypt' feature to NeoMutt, which allows
to securely store encrypted passwords in .neomuttrc and decrypt them
without typing anything, as long as the ssh key used to encrypt the
password is loaded or forwarded by the ssh-agent.

For more information see the modified imap_pass, pop_pass and smtp_pass
documentation in neomuttrc.5

The new feature is disabled by default as it depends on libagentcrypt
which is available at https://github.com/ndilieto/libagentcrypt

To enable it, first install libagentcrypt and then add --agentcrypt to
NeoMutt's ./configure invocation before rebuilding.