Replace `try` variable with `user`
`try` is a keyword in c++

While this should not affect C code, it affects unfortunately the
tooling.

This specific instance triggers a "syntax error" in Cppcheck, so it
could hide other problems.

(Cppcheck version  1.86, code parsed with `--language=c --std=c99`)

Of course it is a bug in Cppcheck, but renaming the variable is the most
pragmatic approach, as the bug in Cppcheck does not manifest itself in
a simple program.