Add envelope's Organization field to hcache
This ensures using %W in index_format while having
header cache enabled works as expected.

Note that as we don't modify the header files that
hcachever.sh is looking for, but the serial_dump_envelope
and serial_restore_envelope functions instead, we have
to increment BASEVERSION to ensure that cache built with
previous versions gets invalidated in all cases.

Fixes #1808