fix read-only mbox
If an mbox Mailbox doesn't have write privs, try opening it read-only.

Before the architectural changes, there could be multiple open Mailbox
objects representing a single folder.

By default a folder would be given a read-only Mailbox.  When the folder
needed to be written to, another Mailbox would be opened read-write.
This lead to synchronisation problems between the two Mailboxes.

Co-authored-by: Pietro Cerutti <gahr@gahr.ch>