eliminate MuttIndexWindow from mutt_copy_message()
`mutt_copy_message()` uses a huge tree of functions to process and copy
the header and body of a message.

Some of these functions perform wrapping and some of the wrapping
functions use `MuttIndexWindow` to get the width of the current panel.

These functions needed an extra parameter:

- `mutt_copy_hdr()`
- `mutt_copy_header()`
- `mutt_copy_message()`
- `mutt_copy_message_fp()`
- `mutt_write_one_header()`

The rest used an extra member in `struct State`.
For example, this sequence of functions:

- `mutt_body_handler()`
- `valid_pgp_encrypted_handler()`
- `run_decode_and_handler()`
- `crypt_pgp_encrypted_handler()`
- `pgp_gpgme_encrypted_handler()`
- `mutt_protected_headers_handler()`