fix: abort entire-thread if database lacks message
If the notmuch database is missing a message, abort <entire-thread>
before causing a segfault. When <entire-thread> executes on a message
missing from the database, it segfaults. This happens because the
vfolder returned is empty and not in an properly initialized state for
the notmuch-backend. As such, the notmuch-backend segfaults.

To prevent a segfault, display a warning and pre-maturely exit
<entire-thread>. While the user is left in an empty vfolder, this is
consistent behavior for <entire-thread>. Ordinarily, you're left with
just the notmuch thread, which doesn't exist in this case.

The location of the abort is perfect for adding behavior to handle the
missing message. In the future, we can automatically run 'notmuch new'
after encountering a missing message. That is outside the scope of this
commit.

Fixes #1861