Remove input buffer flush on sigwinch in index
I recall seeing a message on IRC, where someone reported their macro
suddenly aborting after invoking an editor or something similar.  At
the time, I wasn't sure of the problem, but this seems a likely
culprit, if the event caused a sigwinch.

Neither the menu.c nor the pager.c SigWinch handlers do such a thing.
The line was part of the initial import 21 years ago, so there is no
justification given.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/202ec2c8a4bc8ddc7a3ecafbe82af6b167fc6ae2
Co-authored-by: Richard Russon <rich@flatcap.org>