feat: add 'sidebar_unread' color
With the change to 'sidebar_new' to exclude unread counts, there is no
mechanism to highlight mailboxes with unread mail in the sidebar. This
will appear as a regression to those using 'sidebar_new' for such
purpose.

To span the gap, introduce 'sidebar_unread' which restores the
highlighting for unread mailboxes. It's priority is higher than
'sidebar_flagged'.

Closes #1944
Closes #1875