imap: fix crash when new mail arrives
The original code took a copy of the email array, sized exactly for the
number of emails.  If a new mail arrived during the sync process, the
array would overflow causing a crash.

The size of the email array is actually email_max.  This gives us a
little room for new mail.

Ideally read_headers_fetch_new() would resize the email array.

Either way leads to problems sync-ing the original array and notifying
the caller about the new mail.