Try to infer the signing key from the From address (#2035)
* Try to infer the signing key from the From address

Currently, when running without autocrypt support, the signing key is
specified statically via settings such as pgp_sign_as / smime_sign_as
and pgp_default_key / smime_default_key. This makes it impractical to
select a signing key based on the address that is being used to send the
email. This is especially true when the From header is set manually via
edit_headers or in the compose screen.

This commit changes the ncrypt API to accept an AddressList as part of
the signing process. This AddressList consists of the addresses in the
From header and is used to help locate a signing key, if none can be
found using the static settings.

Note that only the GPGMe backend supports this new functionality. I am
not touching the GPG backend.