Turn off auto-clear outside of autocrypt initialization
The auto-clearing code was added in commit 01bc088c, for autocrypt
initial prompting.  It removed having to keep track of every place a
browser or other menu might be displayed and having to remember to
clear it out.

However, clearing when mutt exits is a change of behavior for those
who have turned off alternative screens.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/8914a1ba38ef394db3d89ff8d0e54774ca9dca48
Co-authored-by: Richard Russon <rich@flatcap.org>