notmuch: use query as name for vfolder-from-query
To provide a consistent experience for those who want human-readable
names, set the 'vfolder-from-query' mailbox's name to the user's query.

The previous behavior for 'vfolder-from-query' delegated mailbox
creation to lower-level backend-agnostic code, which lacks the ability
to set a name. As such, 'vfolder-from-query' mailboxes displayed the
encoded notmuch URI when the user wanted a human readable name.

Since we know a Mailbox will be always created from a valid
'vfolder-from-query' execution, we can create it at a high-level and
pass it to the lower-level components, which obviates the need for the
lower-level component to create it.

Closes #2033