Fix IMAP flag comparison code
The flag comparison code checks whether an IMAP server supports a given
flag. Because the check is run with substrings to be concatenated in a
final command to IMAP, the input flag to be checked can end with white
space, e.g., the function is called with "\\Seen " instead of \\"Seen".

I failed to notice this subtlety in ff57a18b and changed a
mutt_str_strncasecmp into a mutt_str_strcasecmp. This commit reverts the
change and makes it more explicit how we are comparing strings.

Fixes #2063