Fix rendering of replacement_char when Charset_is_utf8
Use mutt_addwch instead of addch to draw the replacement_char in
pager.c. The line to draw the replacement char was
addch(replacement_char())
However, if Charset_is_utf8 is true, replacement_char returns 0xfffd,
which is outside the range that addch can handle. Use mutt_addwch
instead so that the pager displays the replacement char properly.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/10a35e5dd874e0dfdcee750a2bdd42cf70c4df20
Co-authored-by: Richard Russon <rich@flatcap.org>