libmutt: fix segfault when alias lookup returns NULL
`mutt_alias_lookup` is allowed to return NULL and thus we should try to
handle that case.

I have a macro in my neomutt configuration that is being rendered a
template file contains placeholders like `@notmuchmutt@`. That
placeholder is then replaced with the path to the notmuch-mutt script. I
had a typo in the file so some occurrences were not being replaced
properly. Thus neomutt tried to use the `mutt_alias_lookup` function on
the string (since it contained the at-sign). That function did return
NULL and in the neomutt code we are then dereferencing `al` by using
`TAILQ_EMPTY` thus leading to a crash.

By properly checking if the `mutt_alias_lookup` invocation actually
returned something non-NULL we can work around this.

mx: also check for NULL before accessing the alias_lookup result

The alias lookup can fail here since `mutt_alias_lookup` is allowed to
return `NULL`. Blindly accessing the returned value might lead to a
segfault.