Fix ability to specify user/login in url for smtp
Even though user was correctly extracted from the smtp url, the login
was always read from the config variable, as it was never set before.
Now, immediately after extracting info from the url we copy the smtp
user to the login field if the user was set, but login wasn't.