config: Add $abort_key option
This can be used to change the key used to abort prompts (previously
unchangeable, now per default Ctrl-G). The key can specified in the same
format as bind/macro commands.

Please note that when using <Esc> as the abort key, you may also want to
set the environment variable ESCDELAY to a low value or even 0 which
will reduce the time that ncurses waits to distinguish singular <Esc>
key presses from the start of a terminal escape sequence. The default
time is 1000 milliseconds and thus quite noticeable.