Keep track of the previously opened mailbox
When opening mailboxes in stacked mode, e.g., in recall-message, a new
mailbox is opened without the current one being closed. The IMAP state
data is associated with the newly opened mailbox, but there is no code
that remembers the previous mailbox. Hence, when the stacked mailbox is
closed, the IMAP state data cannot be associated back to the previous
mailbox. However, the Context is still thinking the previous mailbox is
the "active" one. This results in the index being displayed, but all
operations on it failing in the IMAP layer.

This commit adds a member to the IMAP state data so whenever a mailbox
is opened the current one is saved, and whenever a mailbox is closed,
the previous one is restored.

Fixes #2192