Provide a fast and a correct version of the RFC5322 parsing regex
We provide a variation that doesn't take comments into consideration.
This is fast and works most of the time. In the vast majority of cases
we've been able to find, comments are added at the end of the date, in
the form "Wed, 13 Jun 2007 12:34:56 +0100 (CET)". The fast version of
the regex handles this just fine.

The second version we provide takes into account comments wherever the
RFC allows them. It is slower because the regex engine has to backtrack,
but it is more correct.