fix sidebar indentation
Fixes: #2272

This commit overhauls the sidebar indentation rules.
The old rules lead to inconsistent results and a bug where indent was
only applied to the current account.

The old scheme tried hard to compare the current entry with the previous
one to work out how to indent.  This meant it was hard for the user to
predict the results.

The new scheme is simpler.  It only considers a single Mailbox.

---

The starting point is the Mailbox's name.  If missing, the full path
will be used.

If possible shorten the path using $folder
If not, use heuristics to extract the path, e.g. strip "imap://host.com/"

Calculate the depths of the path using $sidebar_delim_chars as separators.

If $sidebar_component_depth is set, reduce depth by this amount.

If depth is great than zero, display $sidebar_indent_string depth times.

Finally display the saved short path