imap: use the connection delimiter if provided
This changes how imap_fix_path understands a delimiter in an IMAP path.
Previously, the first occurrence of any of the characters in
C_ImapDelimChars was taken to be the delimiter for the path. Now, the
server-provided delimiter - if any - always takes precedence.

This makes C_ImapDelimChars a mere fallback.

Fixes #2110