fix(sidebar): don't indent full mailbox paths
If the sidebar is displaying a full mailbox path while using
'sidebar_short_path=yes', it will indent the entry based on the number
of delimiters in the path. Since '/' is a default delimiter, there is
significant indentation for local folders.

To resolve this, if the sidebar displays the full path, ignore the
'sidebar_short_path' for that entry. While full paths aren't
particularly user friendly, indented full paths are worse.