Fix function declaration of mutt_window_get_coords
The function declaration had a different meaning compared to the
function definition, as `row` and `col` where swapped.

As the row and column have the same type, it did not cause a linking
error.
Apparently all the places where `mutt_window_get_coords` has been used,
it has been used consistently, so just changing the definition is
enough.