Don't mess with the original pattern when running IMAP searches
Fixes #2349

By using SLIST and friends we give up the possibility to treat a whole list
and a sub-chunk of the whole list as equal types. Because we now have a
dedicated "head" type for lists, 6edb2a71426f239dbbb0 had to change the pattern

```c
a = a->next;
```

into

```c
SLIST_REMOVE_HEAD(a, entries);
```

The two are different in that the second actually changes `a` for the
caller.

The fix here is to make a copy of the original list. It is hackish, but
I don't see a better way right now.