refactor: mutt_list_str_split()
This function used to return a STAILQ_HEAD, but that hid a subtle bug
when the STAILQ was returned empty.

When the STAILQ is initialised, `stqh_last` is set to the address of
`stqh_first` which equates to the address of the STAILQ itself.
The STAILQ is on the stack.

If the function returned without adding anything to the STAILQ, the
`stqh_last` pointer would now be invalid, pointing to the now-dead stack.