refactor: rfc2231_encode_string()
This function used to return a TAILQ_HEAD, but that hid a subtle bug
when the TAILQ was returned empty.

When the TAILQ is initialised, `tqh_last` is set to the address of
`tqh_first` which equates to the address of the TAILQ itself.
The TAILQ is on the stack.

If the function returned without adding anything to the TAILQ, the
`tqh_last` pointer would now be invalid, pointing to the now-dead stack.