refactor: generate_body()
Make mutt_inline_forward() usage consistent in generate_body().

generate_body() works on a potential list of mails but mutt_inline_forward
was taking an optional pointer on a mail and could list itself the
mailbox. Now it takes a non-optional pointer and the generate_body() is
reponsible for calling the function as many time as it has mails tagged.

It also fixed the fact that one could not tag several messages and
forward them at once via `;f`

We now work directly on the EmailList instead of making a difference
whether one or several mails are passed.