Remove $ssl_starttls check for IMAP PREAUTH
Checking $ssl_starttls provides no real protection, because an
attacker can just as easily spoof "* OK" and strip the STARTTLS
capability as it can spoof "* PREAUTH".  The only way to really
protect again the MITM is through $ssl_force_tls.

Add documentation about STARTTLS, $tunnel, and the current PREAUTH
exception when using $tunnel.

The behavior of NeoMutt about $tunnel is somewhat inconsistent: is it
considered secure or not?  For PREAUTH, to avoid breaking
configurations, we assume it is secure.  But at the same time, NeoMutt is
still negotiating STARTTLS for other $tunnel connections.

This will be resolved in master for the next release; probably by
adding a $tunnel_is_secure config variable defaulting "yes" and
removing the STARTTLS negotiation in that case.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/e37516c3aeb11668bd8e262dfa39cb568c0702c0
Co-authored-by: Pietro Cerutti <gahr@gahr.ch>