Add $tunnel_is_secure config, defaulting to true
The config variable is to resolve an ambiguity in NeoMutt about whether
using $tunnel is secure.

On the one hand, the examples in the manual show using ssh or a direct
pipe to a program.  Many users do this to connect to an IMAP server
with PREAUTH configured, relying on the tunnel to be secured by ssh or
by the fact that it's a local pipe.

On the other hand, the NeoMutt connection code still respects
$ssl_starttls and $ssl_force_tls, as if the $tunnel connection were
not already secured.

After some discussion on mutt-dev, it seemed the best idea to assume
the connection is secure by default, in order to not break IMAP
PREAUTH connections, but to provide a configuration variable in case
there are situations where it is not.

Thanks to Aaron Schrab for the original idea of setting conn->ssf for
$tunnel in his patch to ticket 250.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/c9b58e3df046c60417720b4054ffbb2422360e0f
Co-authored-by: Pietro Cerutti <gahr@gahr.ch>