fix(sidebar): abbreviate/shorten what user sees
If a user opted to display a name, perform the sidebar shorten,
abbreviation, and indentation on the name, not the path. Otherwise,
sidebar may render names incorrectly due to operation being calculated
on the path. Notmuch users are more susceptible since queries often
contain file paths or other characters such periods.

If a user wants to display a name but operate under a folder hierarchy,
they must represent that hierarchy in the name.