notify: perform garbage-collection
Delay removing items from the list of Observers and garbage-collect them
later.

A useful feature of the notifications is to use them for unregistering
an Observer.  Unfortunately, this can cause problems because both
`send()` and `notify_observer_remove()` both iterating through the same
`STAILQ`.

`send()` uses `STAILQ_FOREACH_SAFE()` which stores the 'next' item in
the list.  If that next item is deleted, then `send()` crashes.

The solution is to *empty* the `ObserverNode` in `notify_observer_remove()`
and then *remove* it after `send()` is complete.