Fix index tracking functionality
We used to set the current email too early: first we need to go off to
mbox_check to fetch any changes and update index accordingly, then we
can update the current email.

Also, I don't see a reason why changes to the mailbox should not be
propagated when in attach msg mode. Failing to do so might cause a crash
if the mailbox changes under the hood.