Use the email's IMAP UID instead of an increasing number as index
Index is used for [reverse-|last-]mailbox-order. Most backends assign an
increasing number to each Email as it's created parsing the mailbox. For
IMAP, we have the luxury of UIDs assigned for us by the server, see
https://tools.ietf.org/html/rfc3501#section-2.3.1.1

This saves us from reassigning ascending numbers on expunge, which was
the cause for the bug.

Fixes #2651