Reenable $ssl_force_tls
The next release will be 2.0, providing more justification for
(reasonable) breaking changes.  In 2020, this should be the default.

The documentation seems to already make appropriate warnings about
$tunnel_is_secure.  Since that option is introduced this release,
there is no need to worry about existing users with it unset.

However, there were a couple stray double-quotes in the documentation
for that, so clean those up in this commit.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/9204b24e99767ae06b5df25eca55c028d702528b

Co-authored-by: Kevin McCarthy <kevin@8t8.us>