feat(notmuch): show originating folder in index
Changes the behavior of '%b' in 'index_format' to display the email's
folder instead of the notmuch query if it exists. In a notmuch mailbox,
the current behavior of index_format's %b' expando displays the query
portion of a notmuch mailbox declaration.

The reference documentation for 'index_format' says '%b' shows "Filename
of the original message folder (think mailbox)". By my interpretation,
displaying the notmuch query is incorrect as it _is not_ the original
message folder. The underlying maildir folder _is_ the original message
folder. Therefore it should be shown instead.

To improve displayed information, have the folder path be starting at
the same level as the notmuch database folder. Without this extra
information, it's easy for the index to show ambiguous information.

For example, assume there are two underlying maildirs account1/INBOX and
account2/INBOX and a notmuch mailbox with query "tag:inbox", the index
may display:

INBOX
INBOX
INBOX
...

It's impossible to distinguish which INBOX the index means. With the new
implementation, this is displayed:

account1/INBOX
account1/INBOX
account2/INBOX

This solves the ambiguity problem.

[0] https://neomutt.org/guide/reference.html#3-169-%C2%A0index_format