Fix detection of mbox files with new mail
mbox_mbox_check_stats() is supposed to return the number of new
messages, but if flags = 0 we only look at mtime and atime values
without actually parsing the mbox file, so we don't know how many new
messages there are.  We just know there are _some_, so let's
return 1 and hope for the best.

Closes #2740.