Ensure mutt_extract_token() never returns a NULL dest->data
Adopted from mutt:
Commit e5a32a61 removed a 'mutt_buffer_addch (dest, 0)' at the end of
the function.  Most callers had been converted to use the buffer pool,
and the call was strange since buffers self-terminate.

However, this line covered up logic errors in some of the callers,
which assumed the buffer->data could not be NULL afterwards.

I will try to fix up callers with the logic errors in master.  This is
to fix the problem in stable, and also ensure future callers don't
make the same mistake.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/f420be68d517268f5e8600bc01808c3ae4b5b6ba