Fix hdr_order to use the longest match
Previously, NeoMutt would find the first match and use that.

However, the example shown in the manual:
  hdr_order From Date: From: To: Cc: Subject:
gives the expectation that an mbox From_ line would print first, then
the Date:, and then the From: header.

Change the matcher to scan all HeaderOrderList entries and use the
longest match.

Update the documentation to make it clear From_ lines are printed out
with other "real" headers, and can be manipulated via ignore and
hdr_order commands.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/3d6e7df769841f34227a9065401c193b756efde2

Co-authored-by: Kevin McCarthy <kevin@8t8.us>