mutt_save_message refactoring: combine decode & decrypt
decode and decrypt arguments for mutt_save_message and
mutt_save_message_ctx are mutually exclusive, e.g. the following is
always true:

    (decode && decrypt) == false

This patch introduces an enum MessageTransformOpt that is a tri-state of
possible transformation options:
- none
- decode
- decrypt

This reduces the number of requried arguments for these functions and
makes invalid state( both decode and decrypt set to true)
un-representable.